requirement.txt
google-generativeai
speechrecognition
pyttsx3
pillow
argparse
shlex
logging
